adding autodetect for nvidia

+++ Scripts/restore_sgz.sh
@@ -18,12 +18,16 @@ if [ `grep "Current=corners" /etc/sddm.conf.d/kde_settings.conf | wc -w` -eq 0 ]
 fi
 
 # grub
-if [ ! -f /etc/default/grub.t2.bkp ]
+if [ ! -f /etc/default/grub.t2.bkp ] && [ ! -f /boot/grub/grub.t2.bkp ] 
     then
     sudo cp /etc/default/grub /etc/default/grub.t2.bkp
 
+    if [ `lspci -k | grep -A 2 -E "(VGA|3D)" | grep -i nvidia | wc -l` -gt 0 ]
+        then
+        sudo sed -i "/^GRUB_CMDLINE_LINUX_DEFAULT=/c\GRUB_CMDLINE_LINUX_DEFAULT=\"loglevel=3 quiet splash nvidia_drm.modeset=1\"" /etc/default/grub
+    fi
+
     sudo sed -i "/^GRUB_DEFAULT=/c\GRUB_DEFAULT=saved
-    /^GRUB_CMDLINE_LINUX_DEFAULT=/c\GRUB_CMDLINE_LINUX_DEFAULT=\"loglevel=3 quiet splash nvidia_drm.modeset=1\"
     /^GRUB_GFXMODE=/c\GRUB_GFXMODE=1280x1024x32,auto
     /^#GRUB_THEME=/c\GRUB_THEME=\"/usr/share/grub/themes/pochita/theme.txt\"
     /^#GRUB_SAVEDEFAULT=true/c\GRUB_SAVEDEFAULT=true" /etc/default/grub
